% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf_dashboard.R
\docType{package}
\name{dcf.dashboard}
\alias{dcf.dashboard}
\alias{dcf.dashboard-package}
\title{A generic shiny app to report data availability from dcf-shiny databases}
\description{
A generic shiny app to report data availability from dcf-shiny databases
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdiwg/dcf-dashboard}
  \item Report bugs at \url{https://github.com/fdiwg/dcf-dashboard/issues}
}

}
\author{
\strong{Maintainer}: Emmanuel Blondel \email{emmanuel.blondel1@gmail.com} (\href{https://orcid.org/0000-0002-5870-5762}{ORCID})

Authors:
\itemize{
  \item Alexandre Bennici (\href{https://orcid.org/0000-0003-2160-3487}{ORCID})
}

}
